// src/components/models/inventoryModel.tsx
// export interface Todo {
//     id: number;
//     userId: number;
//     title: string;
//     completed: boolean;
//   }
  export interface Article{
    code: string;
    description: string;
    price: number;
    id : number;
    companyId : number;
    active: boolean;
    createDate: Date;
    updateDate: Date;
    total : number;
    brand : string;
    strGroup: string;
    quantity: number;
}